+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.4.2
+-----
+- Added manufacturer property to product model
+- Changed exception message in error handler
+- Fixed identity linker is type bug
+
 2.4.1
 -----
 - Added compatibility method to config class